+++ controllers/front/callbackfail.php
@@ -43,7 +43,6 @@ class AltapayCallbackfailModuleFrontController extends ModuleFrontController
         $lockFileName = sys_get_temp_dir() . DIRECTORY_SEPARATOR . 'callback_lock_' . md5($postData['transaction_id']) . '.lock';
         $lockFileHandle = lockCallback($lockFileName);
 
-        $message = '';
         $customerID = $this->context->customer->id;
         $orderStatus = (int) Configuration::get('authorized_payments_status');
         if (empty($orderStatus)) {
@@ -51,6 +50,7 @@ class AltapayCallbackfailModuleFrontController extends ModuleFrontController
         }
 
         $callback = new API\PHP\Altapay\Api\Ecommerce\Callback($postData);
+        $transaction = null;
         try {
             $response = $callback->call();
             $shopOrderId = $response->shopOrderId;
@@ -118,14 +118,8 @@ class AltapayCallbackfailModuleFrontController extends ModuleFrontController
                         and $altapay_order_details[0]['paymentStatus'] === 'succeeded'
                         and $altapay_order_details[0]['payment_id'] != $transactionID
                         and ($postData['status'] === 'succeeded' or $transaction->ReservedAmount > 0)) {
-                        //refund or release incoming payment request
-                        if (in_array($transaction->TransactionStatus, ['captured', 'bank_payment_finalized'], true)) {
-                            $api = new API\PHP\Altapay\Api\Payments\RefundCapturedReservation(getAuth());
-                        } else {
-                            $api = new API\PHP\Altapay\Api\Payments\ReleaseReservation(getAuth());
-                        }
-                        $api->setTransaction($transactionID);
-                        $api->call();
+                        // Refund or Release incoming payment request
+                        refundOrReleaseTransactionByStatus($transaction);
                         unlockCallback($lockFileName, $lockFileHandle);
                         Tools::redirect('index.php?controller=order-confirmation&id_cart=' . (int) $cart->id . '&id_module=' . (int) $this->module->id . '&id_order=' . (int) $order_id . '&key=' . $customer->secure_key);
                     }
@@ -157,7 +151,7 @@ class AltapayCallbackfailModuleFrontController extends ModuleFrontController
                         $response = capturePayment($order->id, $transactionID, $amountPaid);
                         $orderStatusCaptured = (int) Configuration::get('PS_OS_PAYMENT');
                         if ($orderStatusCaptured != $orderStatus) {
-                            $order->setCurrentState($orderStatusCaptured);
+                            setOrderStateIfNotExistInHistory($order, $orderStatusCaptured);
                         }
                     }
 
@@ -192,9 +186,18 @@ class AltapayCallbackfailModuleFrontController extends ModuleFrontController
             $errorMessage = $e->getMessage();
         }
 
+        // Successful order exists, set its status to cancel
+        $order = getOrderFromUniqueId($postData['shop_orderid']);
+        if (Validate::isLoadedObject($order) and $transaction->ReservedAmount == 0) {
+            updatePaymentStatus($postData['transaction_id'], $postData['status']);
+            saveLastErrorMessage($postData['transaction_id'], $errorMessage);
+            $orderStatusCancelled = (int) Configuration::get('PS_OS_CANCELED');
+            setOrderStateIfNotExistInHistory($order, $orderStatusCancelled);
+        }
+
         $status = strtolower($response->Result);
         if ($status === 'cancelled') {
-            $unique_id = Tools::getValue('shop_orderid');
+            $unique_id = $postData['shop_orderid'];
             // Updated transaction record to cancel
             $pI = pSQL($unique_id);
             $q = 'UPDATE `' . _DB_PREFIX_ . 'altapay_transaction` set `is_cancelled`=1, `transaction_status` = "' . $status . '" WHERE `unique_id`=\'' . $pI . '\'';